Create peerReview
This was an application that allowed physicians to access their peerReview folders from any computer connected to our network. This is the currently working version. I have another update that I hope to push out that contains proper logging information and some stability fixes.

This is a gui made with PowerShell Forms from Sapien Powershell studio.